Pages are the WORST!
When people want to get mean, they get anonymous.  Fake profiles violate Facebook TOS and can therefore be removed.  However, some users take it a step further and create a "Page" and use that to be pretty disruptive without facing disciplinary recourse from Facebook.  Would it be appropriate to create a new standard that limits engagement to profiles only, and disallowing "pages"?